fix: reconcile model invoke with latest main
