+++ pkg/services/models/transports/cli/composition_test.go
@@ -8,7 +8,6 @@ import (
 	"fmt"
 	"io"
 	"net/http"
-	"net/http/httptest"
 	"os"
 	"path/filepath"
 	"strings"
@@ -568,44 +567,6 @@ func TestNewInvokesThroughCompositionProviderOwnedPath(t *testing.T) {
 	}
 }
 
-func TestNew_ServerValidationUsesHTTPFallbackWhenCompositionRootExists(t *testing.T) {
-	t.Parallel()
-
-	server := httptest.NewServer(http.HandlerFunc(func(writer http.ResponseWriter, request *http.Request) {
-		if request.URL.Path != "/models/known" {
-			http.NotFound(writer, request)
-			return
-		}
-		writer.Header().Set("Content-Type", "application/json")
-		_, _ = io.WriteString(writer, `{"name":"known","operations":[{"name":"TTS"}]}`)
-	}))
-	t.Cleanup(server.Close)
-
-	localCalled := false
-	invocation := factorySessionPresentationInvocation{
-		root: compositionModelsRoot{
-			getModel: func(context.Context, string) (modelinference.Detail, error) {
-				localCalled = true
-				return modelinference.Detail{}, nil
-			},
-		},
-	}
-	service := modelscli.New(compositionHTTPProtocol(t), invocation)
-	var out bytes.Buffer
-	if err := service.Invoke(modelscli.InvokeConfig{
-		Context: context.Background(), ModelName: "known", Operation: "TTS", Text: "hello",
-		Server: server.URL, JSON: true, Output: &out,
-	}); err != nil {
-		t.Fatalf("Invoke() error = %v", err)
-	}
-	if localCalled {
-		t.Fatal("server-bound validation opened the locally composed Models catalog")
-	}
-	if !strings.Contains(out.String(), `"mode":"VALIDATION_ONLY"`) {
-		t.Fatalf("Invoke() output = %q, want validation-only metadata", out.String())
-	}
-}
-
 func TestRootAdapter_InvokeGenericSingleOutputWritesOnlyPayload(t *testing.T) {
 	t.Parallel()
 
@@ -673,55 +634,6 @@ func TestRootAdapter_InvokeGenericMultipleOutputsRejectsBeforeRoot(t *testing.T)
 	}
 	if called {
 		t.Fatal("generic root was called after multi-output preflight rejection")
-	}
-}
-
-func TestRootAdapter_InvokeGenericJSONIsValidationOnly(t *testing.T) {
-	t.Parallel()
-
-	scope := testRuntimeScope(t)
-	artifact := testArtifactRef(t, "artifact:usage")
-	service := modelscli.NewService(modelscli.Config{
-		Models: stubModelsRoot{
-			getCatalogModel: func(context.Context, modelinference.GetModelRequest) (modelinference.GetModelResult, error) {
-				return genericCLIModel("omni", modelinference.OperationOMNI,
-					modelinference.OperationSlot{Name: "text", Modality: modelinference.ModalityText},
-					modelinference.OperationSlot{Name: "usage", Modality: modelinference.ModalityJSON}), nil
-			},
-			invokeModel: func(context.Context, modelinference.InvokeModelRequest) (modelinference.InvokeModelResult, error) {
-				return modelinference.InvokeModelResult{Outputs: []modelinference.InferenceOutput{
-					{Name: "text", Modality: modelinference.ModalityText, Content: "answer"},
-					{Name: "usage", Modality: modelinference.ModalityJSON, Artifact: &modelinference.InferenceArtifact{
-						Artifact: artifact, MediaType: "application/json", SizeBytes: 7,
-					}},
-				}}, nil
-			},
-		},
-		OpenInvokeScope: func(context.Context, modelscli.InvokeConfig) (modelscli.InvokeRuntimeScope, error) {
-			return modelscli.InvokeRuntimeScope{Scope: scope}, nil
-		},
-	})
-
-	var out bytes.Buffer
-	if err := service.Invoke(modelscli.InvokeConfig{
-		Context: context.Background(), ModelName: "omni", Operation: modelinference.OperationOMNI,
-		Text: "hello", JSON: true, Output: &out,
-	}); err != nil {
-		t.Fatalf("Invoke() error = %v", err)
-	}
-	var response struct {
-		ModelName         string `json:"modelName"`
-		Operation         string `json:"operation"`
-		Mode              string `json:"mode"`
-		ValidationOnly    bool   `json:"validationOnly"`
-		InferenceExecuted bool   `json:"inferenceExecuted"`
-	}
-	if err := json.Unmarshal(out.Bytes(), &response); err != nil {
-		t.Fatalf("decode validation JSON: %v\n%s", err, out.String())
-	}
-	if response.ModelName != "omni" || response.Operation != modelinference.OperationOMNI ||
-		response.Mode != "VALIDATION_ONLY" || !response.ValidationOnly || response.InferenceExecuted {
-		t.Fatalf("validation response = %#v, want validation-only metadata", response)
 	}
 }
