+++ pkg/services/models/transports/cli/models.go
@@ -89,7 +89,8 @@ type modelInvocationValidationResponse struct {
 }
 
 func validationOnlyModelInvoke(cfg InvokeConfig) bool {
-	return cfg.JSON && strings.TrimSpace(cfg.OutputPath) == "" && len(cfg.OutputMappings) == 0
+	return cfg.JSON && strings.TrimSpace(cfg.OutputPath) == "" &&
+		len(cfg.InputMappings) == 0 && len(cfg.OutputMappings) == 0
 }
 
 func writeValidationOnlyModelInvokeResponse(output io.Writer, modelName, operation string) error {
